Frontend add GraphPage.tsx so the filter is not updated when Graph.tsx updates
